+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## NEXT
+
+- Update deps
+
 ## 0.4.0
 
 - Fix typing issue; all generic `TMessage` args now inherit from MongoDB's `Document` interface